stashing changes

+++ stick.jl
@@ -31,13 +31,13 @@ function getSalt()::UInt128
 end
 
 function main()
-    chainId::UInt8 = 1
+    chainId = 1
     entropy = 0x949f147e5fb733f43d887bd3f9455dad768029a75baa8c63ec2d749439935d59
     gemAddr::String = "0xC67DED0eC78b849e17771b2E8a7e303B4dAd6dD4"
     userAddr::String = "0xFe9C7c2BD25d91B4f2E9C6AE83C11e1465eC63F0"
-    kind::UInt8 = 0
-    nonce::UInt8 = 0
-    diff::Float64 = 1
+    kind = 0
+    nonce = 0
+    diff = 1
 
     target = ^(2, 256) / diff
